Move event handlers out of WorkerChannel
