+++ test/eventHandlers/TestEventStream.ts
@@ -3,8 +3,8 @@
 
 import { EventEmitter } from 'events';
 import * as sinon from 'sinon';
-import { AzureFunctionsRpcMessages as rpc } from '../azure-functions-language-worker-protobuf/src/rpc';
-import { IEventStream } from '../src/GrpcClient';
+import { AzureFunctionsRpcMessages as rpc } from '../../azure-functions-language-worker-protobuf/src/rpc';
+import { IEventStream } from '../../src/GrpcClient';
 
 export class TestEventStream extends EventEmitter implements IEventStream {
     written: sinon.SinonSpy;
