+++ test/eventHandlers/invocationRequest.test.ts
@@ -4,18 +4,23 @@
 import { expect } from 'chai';
 import 'mocha';
 import * as sinon from 'sinon';
-import { AzureFunctionsRpcMessages as rpc } from '../azure-functions-language-worker-protobuf/src/rpc';
-import { FunctionInfo } from '../src/FunctionInfo';
-import { FunctionLoader } from '../src/FunctionLoader';
-import { WorkerChannel } from '../src/WorkerChannel';
+import { AzureFunctionsRpcMessages as rpc } from '../../azure-functions-language-worker-protobuf/src/rpc';
+import { FunctionInfo } from '../../src/FunctionInfo';
+import { FunctionLoader } from '../../src/FunctionLoader';
+import { WorkerChannel } from '../../src/WorkerChannel';
+import { beforeEventHandlerTest } from './beforeEventHandlerTest';
 import { TestEventStream } from './TestEventStream';
 import LogCategory = rpc.RpcLog.RpcLogCategory;
 import LogLevel = rpc.RpcLog.Level;
 
-describe('WorkerChannel', () => {
+describe('invocationRequest', () => {
     let channel: WorkerChannel;
     let stream: TestEventStream;
     let loader: sinon.SinonStubbedInstance<FunctionLoader>;
+
+    beforeEach(() => {
+        ({ stream, loader, channel } = beforeEventHandlerTest());
+    });
 
     const sendInvokeMessage = (
         inputData?: rpc.IParameterBinding[] | null,
@@ -139,234 +144,6 @@ describe('WorkerChannel', () => {
             },
         },
     };
-
-    beforeEach(() => {
-        stream = new TestEventStream();
-        loader = sinon.createStubInstance<FunctionLoader>(FunctionLoader);
-        channel = new WorkerChannel('workerId', stream, loader);
-    });
-
-    it('responds to init', () => {
-        const initMessage = {
-            requestId: 'id',
-            workerInitRequest: {
-                capabilities: {},
-            },
-        };
-
-        const expectedOutput = {
-            requestId: 'id',
-            workerInitResponse: {
-                capabilities: {
-                    RpcHttpBodyOnly: 'true',
-                    RpcHttpTriggerMetadataRemoved: 'true',
-                    IgnoreEmptyValuedRpcHttpHeaders: 'true',
-                    UseNullableValueDictionaryForHttp: 'true',
-                },
-                result: {
-                    status: rpc.StatusResult.Status.Success,
-                },
-            },
-        };
-
-        expectedOutput.workerInitResponse.capabilities['TypedDataCollection'] = 'true';
-
-        stream.addTestMessage(initMessage);
-        sinon.assert.calledWith(stream.written);
-    });
-
-    it('does not init for Node.js v8.x and v2 compatability = false', () => {
-        const version = process.version;
-        if (version.split('.')[0] === 'v8') {
-            const initMessage = {
-                requestId: 'id',
-                workerInitRequest: {
-                    capabilities: {},
-                },
-            };
-
-            expect(() => stream.addTestMessage(initMessage)).to.throw(
-                `Incompatible Node.js version (${process.version}). The version of the Azure Functions runtime you are using (v3) supports Node.js v10.x and v12.x. Refer to our documentation to see the Node.js versions supported by each version of Azure Functions: https://aka.ms/functions-node-versions`
-            );
-        }
-    });
-
-    it('responds to function load', async () => {
-        stream.addTestMessage({
-            requestId: 'id',
-            functionLoadRequest: {
-                functionId: 'funcId',
-                metadata: {},
-            },
-        });
-        // Set slight delay
-        await new Promise((resolve) => setTimeout(resolve, 100));
-        sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-            requestId: 'id',
-            functionLoadResponse: {
-                functionId: 'funcId',
-                result: {
-                    status: rpc.StatusResult.Status.Success,
-                },
-            },
-        });
-    });
-
-    it('handles function load exception', () => {
-        const err = new Error('Function throws error');
-        err.stack = '<STACKTRACE>';
-
-        loader.load = sinon.stub().throws(err);
-        channel = new WorkerChannel('workerId', stream, loader);
-        stream.addTestMessage({
-            requestId: 'id',
-            functionLoadRequest: {
-                functionId: 'funcId',
-                metadata: {},
-            },
-        });
-        sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-            requestId: 'id',
-            functionLoadResponse: {
-                functionId: 'funcId',
-                result: {
-                    status: rpc.StatusResult.Status.Failure,
-                    exception: {
-                        message: "Worker was unable to load function undefined: 'Error: Function throws error'",
-                        stackTrace: '<STACKTRACE>',
-                    },
-                },
-            },
-        });
-    });
-
-    it('reloads environment variables', () => {
-        process.env.PlaceholderVariable = 'TRUE';
-        stream.addTestMessage({
-            requestId: 'id',
-            functionEnvironmentReloadRequest: {
-                environmentVariables: {
-                    hello: 'world',
-                    SystemDrive: 'Q:',
-                },
-                functionAppDirectory: null,
-            },
-        });
-        sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-            requestId: 'id',
-            functionEnvironmentReloadResponse: {
-                result: {
-                    status: rpc.StatusResult.Status.Success,
-                },
-            },
-        });
-        expect(process.env.hello).to.equal('world');
-        expect(process.env.SystemDrive).to.equal('Q:');
-        expect(process.env.PlaceholderVariable).to.be.undefined;
-    });
-
-    it('reloading environment variables removes existing environment variables', () => {
-        process.env.PlaceholderVariable = 'TRUE';
-        process.env.NODE_ENV = 'Debug';
-        stream.addTestMessage({
-            requestId: 'id',
-            functionEnvironmentReloadRequest: {
-                environmentVariables: {},
-                functionAppDirectory: null,
-            },
-        });
-        sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-            requestId: 'id',
-            functionEnvironmentReloadResponse: {
-                result: {
-                    status: rpc.StatusResult.Status.Success,
-                },
-            },
-        });
-        expect(process.env).to.be.empty;
-    });
-
-    it('reloads empty environment variables without throwing', () => {
-        expect(() => {
-            stream.write({
-                requestId: 'id',
-                functionEnvironmentReloadRequest: {
-                    environmentVariables: {},
-                    functionAppDirectory: null,
-                },
-            });
-        }).to.not.throw();
-
-        expect(() => {
-            stream.write({
-                requestId: 'id',
-                functionEnvironmentReloadRequest: null,
-            });
-        }).to.not.throw();
-
-        expect(() => {
-            stream.write({
-                requestId: 'id',
-                functionEnvironmentReloadRequest: {
-                    environmentVariables: null,
-                    functionAppDirectory: null,
-                },
-            });
-        }).to.not.throw();
-    });
-
-    it('reloads environment variable and keeps cwd without functionAppDirectory', () => {
-        const cwd = process.cwd();
-        stream.addTestMessage({
-            requestId: 'id',
-            functionEnvironmentReloadRequest: {
-                environmentVariables: {
-                    hello: 'world',
-                    SystemDrive: 'Q:',
-                },
-                functionAppDirectory: null,
-            },
-        });
-        sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-            requestId: 'id',
-            functionEnvironmentReloadResponse: {
-                result: {
-                    status: rpc.StatusResult.Status.Success,
-                },
-            },
-        });
-        expect(process.env.hello).to.equal('world');
-        expect(process.env.SystemDrive).to.equal('Q:');
-        expect(process.cwd() == cwd);
-    });
-
-    it('reloads environment variable and changes functionAppDirectory', () => {
-        const cwd = process.cwd();
-        const newDir = '/';
-        stream.addTestMessage({
-            requestId: 'id',
-            functionEnvironmentReloadRequest: {
-                environmentVariables: {
-                    hello: 'world',
-                    SystemDrive: 'Q:',
-                },
-                functionAppDirectory: newDir,
-            },
-        });
-        sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-            requestId: 'id',
-            functionEnvironmentReloadResponse: {
-                result: {
-                    status: rpc.StatusResult.Status.Success,
-                },
-            },
-        });
-        expect(process.env.hello).to.equal('world');
-        expect(process.env.SystemDrive).to.equal('Q:');
-        expect(process.cwd() != newDir);
-        expect(process.cwd() == newDir);
-        process.chdir(cwd);
-    });
 
     it('invokes function', () => {
         loader.getFunc.returns((context) => context.done());
@@ -621,77 +398,44 @@ describe('WorkerChannel', () => {
 
             sendInvokeMessage([httpInputData], getHttpTriggerDataMock());
         });
+    });
 
-        it('responds to worker status', async () => {
-            stream.addTestMessage({
-                requestId: 'id',
-                workerStatusRequest: {},
-            });
-            // Set slight delay
-            await new Promise((resolve) => setTimeout(resolve, 100));
-            sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-                requestId: 'id',
-                workerStatusResponse: {},
-            });
-        });
+    it('returns and serializes falsy value in Durable: ""', () => {
+        loader.getFunc.returns((context) => context.done(null, ''));
+        loader.getInfo.returns(new FunctionInfo(activityBinding));
 
-        it('returns and serializes falsy value in Durable: ""', () => {
-            loader.getFunc.returns((context) => context.done(null, ''));
-            loader.getInfo.returns(new FunctionInfo(activityBinding));
+        sendInvokeMessage([], getHttpTriggerDataMock());
 
-            sendInvokeMessage([], getHttpTriggerDataMock());
+        const expectedOutput = [];
+        const expectedReturnValue = {
+            string: '',
+        };
+        assertInvocationSuccess(expectedOutput, expectedReturnValue);
+    });
 
-            const expectedOutput = [];
-            const expectedReturnValue = {
-                string: '',
-            };
-            assertInvocationSuccess(expectedOutput, expectedReturnValue);
-        });
+    it('returns and serializes falsy value in Durable: 0', () => {
+        loader.getFunc.returns((context) => context.done(null, 0));
+        loader.getInfo.returns(new FunctionInfo(activityBinding));
 
-        it('returns and serializes falsy value in Durable: 0', () => {
-            loader.getFunc.returns((context) => context.done(null, 0));
-            loader.getInfo.returns(new FunctionInfo(activityBinding));
+        sendInvokeMessage([], getHttpTriggerDataMock());
 
-            sendInvokeMessage([], getHttpTriggerDataMock());
+        const expectedOutput = [];
+        const expectedReturnValue = {
+            int: 0,
+        };
+        assertInvocationSuccess(expectedOutput, expectedReturnValue);
+    });
 
-            const expectedOutput = [];
-            const expectedReturnValue = {
-                int: 0,
-            };
-            assertInvocationSuccess(expectedOutput, expectedReturnValue);
-        });
+    it('returns and serializes falsy value in Durable: false', () => {
+        loader.getFunc.returns((context) => context.done(null, false));
+        loader.getInfo.returns(new FunctionInfo(activityBinding));
 
-        it('returns and serializes falsy value in Durable: false', () => {
-            loader.getFunc.returns((context) => context.done(null, false));
-            loader.getInfo.returns(new FunctionInfo(activityBinding));
+        sendInvokeMessage([], getHttpTriggerDataMock());
 
-            sendInvokeMessage([], getHttpTriggerDataMock());
-
-            const expectedOutput = [];
-            const expectedReturnValue = {
-                json: 'false',
-            };
-            assertInvocationSuccess(expectedOutput, expectedReturnValue);
-        });
-
-        it('logs AzureFiles cold start warning', async () => {
-            process.env.WEBSITE_CONTENTAZUREFILECONNECTIONSTRING = 'test';
-            process.env.WEBSITE_CONTENTSHARE = 'test';
-            process.env.AzureWebJobsScriptRoot = 'test';
-
-            // Accesing private method
-            (channel as any).logColdStartWarning(10);
-
-            // Set slight delay
-            await new Promise((resolve) => setTimeout(resolve, 1000));
-            sinon.assert.calledWith(stream.written, <rpc.IStreamingMessage>{
-                rpcLog: {
-                    message:
-                        'package.json is not found at the root of the Function App in Azure Files - cold start for NodeJs can be affected.',
-                    level: LogLevel.Debug,
-                    logCategory: LogCategory.System,
-                },
-            });
-        });
+        const expectedOutput = [];
+        const expectedReturnValue = {
+            json: 'false',
+        };
+        assertInvocationSuccess(expectedOutput, expectedReturnValue);
     });
 });
